# Optimize use of Eigen::DenseBase::select() for PRelu