MSVC and gcc are both not good at optimizing select(), even in trivial usage outside of ORT.
  gcc seems to do better with -ffast-math (not used by ORT) but /fp:fast does nothing for MSVC
  This PR delivers a 33% speedup on the same model (360us -> 270us on Windows; 205 us -> 153 us on Linux; measured on different systems).

TODO: Examine and fix Elu and other similar activation functions for the use of `Eigen::select`

  Co-authored-by: @fpribeiro

### Description
<!-- Describe your changes. -->



### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->